Refactor code for efficiency
Improved the efficiency of the existing codebase by refactoring components and optimizing performance.
[skip gpt_engineer]
